Daniil's suggestions
Daily Max Leverage
I think it’ll be more correct to expand the range for Daily Max Leverage (Avg) to 30 (now 0-10), otherwise we lose sight of many futures strategies.
Delta Equity 30/45 Days
Increment 5 is too much. In C2Score_AnnReturn we can clearly see that a return of 20% is the most optimal benchmark, and this is less than 2% in 30 days. I propose to experiment with increments 1, 2 and 2.5