New data
New data:
- Daily equity is OPEN of the day for compatibility with Strategy Scoring Workbench now. (It was a closing value of the day.) 
- $33 autotrade fee subtracted from equities.
- Last data are from 2021-06-21 now. Almost 3 months more of data than you have.
- There is about 10% more systems there now. (I guess new systems from last 3 months.)

All that means that we have different account equities than before and the population statistics is slightly different from the previous, but very similar.
Statistics hypotheses tests are slightly different from the same reasons too.

I published new results on Github.
I fixed the first row in the charts. 
But there is a small glitch in the charts there - values on the Y axis should be a half of row upper, and the very first Y value shoud be (usually) zero. 
I will try to fix that, but unfortunately, I do not know how yet...
I guess you will want to look at this new hypothesis tests results. In doubts, a guide is a table in the paragraph "The following table contains numbers of observations in the used statistics intervals."   
   
I mean the first column which contains stats intervals stored in rows like this:
3.8 <= statval < 4  	
3.6 <= statval < 3.8	
....
0.4 <= statval < 0.6	
0.2 <= statval < 0.4	
0 <= statval < 0.2